Initial Smooth Design
I've added some javascript and other assets to make
the website look a bit more smooth and professional.

It now has cards and text that scroll up when the
user enters the website. This is the basic design for the whole
website. The text is a bit wonky in this commit but that's fine.